dispatch-conf: Add support for conf-update.d hook directory
Those hooks can be used by tools that manage /etc to get notified
about updated configuration files. For example, etckeeper could hook
this mechanism like the following:

/etc/portage/conf-update.d/etckeeper
case "${1}" in
    pre-update)
        etckeeper pre-install
    ;;
    post-update)
        etckeeper post-install
    ;;
esac

Currently conf-update.d hooks are called with 4 different events:
- pre-session
- post-session
- pre-update
- post-update

The *-session events are emitted prior starting a new configuration
update sesssion, and when it is finished. That is, the pre-session
event is emitted just before dispatch-conf displays the first
configuration file, and right before it exists.

The *-update events are emitted before and after a configuration file
has been updated. The path of the configuration file is provided as
second hook argument.

Signed-off-by: Florian Schmaus <flo@geekplace.eu>
Closes: https://bugs.gentoo.org/698316
Bug: https://bugs.gentoo.org/260623